Refactored BrowsePets page to fit MVC.
Refactoring of BrowsePets has been completed.
The buttons still need implementation of their filter/order logic of animals that are displayed.

Potentially also need a page system? We can keep it simple and be 1 long page to be scrolled? Future frontend designers need to make this decision.
